+++ src/main.rs
@@ -43,15 +43,6 @@ enum Args {
     },
 
     /// Edit the file to set some data
-    ///
-    /// Use `--write` to actually write the new version back to the file,
-    /// or `--print` to print it to stdout instead.
-    ///
-    /// For legacy reasons, there is no default; either `--write` or `--print`
-    /// is required.  A future version will change the default to `--write`.
-    // Without verbatim_doc_comment, the paragraphs get rewrapped to like
-    // 120 columns wide.
-    #[structopt(verbatim_doc_comment)]
     Set {
         /// Path to the TOML file to edit
         #[structopt(parse(from_os_str))]
@@ -86,11 +77,11 @@ struct GetOpts {
 struct SetOpts {
     /// Print the new version instead of editing the file
     #[structopt(long)]
-    #[allow(dead_code)]
     print: bool,
 
-    /// Write the new version back to the file instead of printing it
+    /// (default) Write the new version back to the file
     #[structopt(long, conflicts_with("print"))]
+    #[allow(dead_code)]
     write: bool,
 }
 
@@ -109,10 +100,6 @@ enum CliError {
 enum SilentError {
     #[error("key not found: {key}")]
     KeyNotFound { key: String },
-
-    // "Silent" because we'll have already printed an error to stderr.
-    #[error("`toml set` requires explicit `--print` or `--write`")]
-    SetNoPrintWrite(),
 }
 
 fn main() {
@@ -258,24 +245,7 @@ fn set(path: &PathBuf, query: &str, value_str: &str, opts: &SetOpts) -> Result<(
     }
     *item = value(value_str);
 
-    if !opts.write && !opts.print {
-        // TODO move this to more the CLI-parsing phase
-        // TODO perhaps make fancier error output
-        eprint!(
-            "------------------------------------------------------------\n\
-             toml: ERROR: `toml set` requires explicit `--print` or `--write`\n\
-             toml:\n\
-             toml: The behavior of plain `toml set` is in transition.\n\
-             toml: Use an explicit `toml set --print` for the old behavior,\n\
-             toml: or `toml set --write` to actually update the TOML file.\n\
-             toml:\n\
-             toml: (In a future version, `--write` will be the default.)\n\
-             ------------------------------------------------------------\n\
-            "
-        );
-        Err(SilentError::SetNoPrintWrite())?;
-    }
-    if !opts.write {
+    if opts.print {
         print!("{}", doc);
     } else {
         fs::write(path, doc.to_string())?;